chain,cozk2p,lib,app,docs: crossing prices, native fees, encrypted reveal
Matching pairs crossing orders. It no longer needs equal prices. The
maker's limit becomes the common execution price; if the maker is a
market order, the taker's limit is used. The buy side gets its price
improvement back as a shielded refund note. A market order carries a
public protection price for the collateral lock and the slippage bound.
Two market orders cannot match, because neither gives an execution
price.

The fee leaves the collateral bank. SendOrder spends collateral-asset
notes for the lock and native `invis` notes for the fee. Each asset
conserves on its own, so the circuit takes four nullifiers and two
change commitments. The order id hashes all four nullifiers.

The reveal is no longer plaintext. RegisterSettleAddr publishes a
signed per-round X25519 key, and the smaller opening travels
ChaCha20-Poly1305 encrypted end to end. Before any opening, both
traders must put a signed pre-open checkpoint on chain. After 10
blocks, a sole uploader can requeue its own order and freeze the
missing peer.

The signing domains move to compare v3 and settle-small/large v2. Both
legs now cover the refund commitment.

The verifying keys, the RQ1-RQ3 results, the design docs and the paper
PDF are regenerated against this protocol. .tmp/ and __pycache__/ are
ignored.

+++ lib/chain/src/note_store.rs
@@ -162,7 +162,9 @@ impl NoteStore {
         if let Some(single) = unspent.iter().rev().find(|r| r.amount >= need) {
             return Some(vec![(*single).clone()]);
         }
-        if unspent.len() >= 2 && unspent[0].amount + unspent[1].amount >= need {
+        if unspent.len() >= 2
+            && u128::from(unspent[0].amount) + u128::from(unspent[1].amount) >= u128::from(need)
+        {
             return Some(vec![unspent[0].clone(), unspent[1].clone()]);
         }
         None
